requirements.txt
numpy
Pillow
scikit-learn
matplotlib
